combine digitizers into /actors/digitizers.py
