--- opengate/actors/DigitizerReadoutActor.py
@@ -1,44 +1,0 @@
-import opengate_core as g4
-from .ActorBase import ActorBase
-from .DigitizerAdderActor import DigitizerAdderActor
-from ..helpers import fatal
-
-
-class DigitizerReadoutActor(g4.GateDigitizerReadoutActor, ActorBase):
-    """
-    This actor is a DigitizerAdderActor + a discretization step:
-    the final position is the center of the volume
-    """
-
-    type_name = "DigitizerReadoutActor"
-
-    @staticmethod
-    def set_default_user_info(user_info):
-        DigitizerAdderActor.set_default_user_info(user_info)
-        user_info.discretize_volume = None
-
-    def __init__(self, user_info):
-        ActorBase.__init__(self, user_info)
-        g4.GateDigitizerReadoutActor.__init__(self, user_info.__dict__)
-        actions = {"StartSimulationAction", "EndSimulationAction"}
-        self.AddActions(actions)
-
-    def __del__(self):
-        pass
-
-    def __str__(self):
-        s = f"DigitizerReadoutActor {self.user_info.name}"
-        return s
-
-    def StartSimulationAction(self):
-        DigitizerAdderActor.set_group_by_depth(self)
-        if self.user_info.discretize_volume is None:
-            fatal(f'Please, set the option "discretize_volume"')
-        depth = self.simulation.volume_manager.get_volume_depth(
-            self.user_info.discretize_volume
-        )
-        self.SetDiscretizeVolumeDepth(depth)
-        g4.GateDigitizerReadoutActor.StartSimulationAction(self)
-
-    def EndSimulationAction(self):
-        g4.GateDigitizerReadoutActor.EndSimulationAction(self)
